feat(proxy): add basic unsplash-proxy
The project uses Cloudflare Workers to proxy API requests to Unsplash.
This way the access key remains a secret and does not have to be exposed
to the web app.

In the future, the proxy can also replay requests for previous images
when there the Unsplash API rate limit is reached.