feat(proxy): add a database for caching photos
Whenever where are issues retrieving photos from the Unsplash API, the
Cloudflare D1 database will be used to retrieve those photos.

This is particularly useful to work around the 50 requests/hour
rate-limit. When the limit is reached, photos will be retrieved from the
D1 database, as opposed to returning errors. The proxy API consumers
will have an easier time getting the photos.

Photos are persisted in the database after each request to the Unsplash
API. In the interest of caching as much as possible, the proxy always
asks for 30 photos, and then stores all of them in the database, even if
not all of them are included in the current response.

Each photo in the database has a `last_used_date` column. This helps
ensure photos are reused as rarely as possible.