refactor(slideshow): use a machine to control the forward slideshow
Create a slideshowForwardMachine to control the forward slideshow.

It only supports displaying images on a timer for now. It asks the
photoFetcherMachine for images.

The `App` component code got much simpler, since it no longer has to
manage the slideshow lifecycle.

It does not support pause/play nor rewind functionalities. Those will be
added in follow-up commits.