feat(slideshow): add rewind slideshow functionality
The rewind functionality is implemented as a separate XState machine
(slideshowRewindMachine), a sibling of slideshowForwardMachine. The
parent slideshowMachine is responsible for controlling the two child
machines.

The App component displays proper UI for these 2 states.

The rewind functionality is currently broken, since it always says there
is nothing in the history. This is because the parent slideshowMachine
does not have access to the latest photoHistory of the
slideshowForwardMachine, and it starts the slideshowRewindMachine with
an empty photo history.