fix(slideshow): initialize history in rewind mode
The previous behavior I saw (not having access to
`slideshowForwardMachine`'s context in the parent `slideshowMachine`)
was caused by the machines being recreated when the state of
`slideshowMachine` was changed from `forward` to `rewind`. This
triggered the `entry` action of `slideshowMachine`, which overwrote the
previous machines with brand new ones, with empty state.

When the `entry` initializing action runs only once, the application
works well.