refactor(slideshow): change photo immediately after resuming
After resuming the slideshow (in either forward or rewind mode), change
the photo immediately, instead of waiting 2 seconds.

The user has more than enough time to look at the photo when it is
paused. When they click the "Resume" button, they want to see something
else as soon as possible. Hence it makes sense to remove the delay.