feat(slideshow): wait for photos to load before starting timer
Problem: in the "forward" mode, sometimes the photo took a long time to
load. The machine started counting the 2s timeout from the moment the
photo started loading, not when it loaded. This made it possible for the
photo to never be shown to the user, because it could take longer than 2
seconds to load. The user would see the previous photo during that time.

Solution: make the machine aware of the photo loading status. Only start
the 2 seconds timer after the photo was loaded.